requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
comm==0.2.2
confection==0.1.4
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.1
contractions==0.1.73
coverage==7.6.0
coveralls==4.0.1
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
et-xmlfile==1.1.0
etils==1.5.2
executing==2.0.1
fastjsonschema==2.19.1
fasttext==0.9.3
fasttext-langdetect==1.0.5
filelock==3.14.0
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.5.0
future==1.0.0
gensim==4.3.2
genson==1.3.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.0
h11==0.14.0
h5py==3.11.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
hypothesis==6.108.2
idna==3.7
ImageHash==4.3.1
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kagglehub==0.2.5
keras==3.3.3
keras-core==0.1.7
keras-cv==0.9.0
keras-nlp==0.12.0
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.7
language_data==1.2.0
llvmlite==0.42.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mock==5.1.0
mpmath==1.3.0
multimethod==1.11.2
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.3.0
plac==1.4.3
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.1.0
pyarrow==17.0.0
pybind11==2.13.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.2.0
pytest-cov==5.0.0
python-calamine==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.6.0
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tabpy==2.10.0
tenacity==8.3.0
tensorflow-datasets==4.9.4
tensorflow-metadata==1.15.0
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
textsearch==0.0.24
thinc==8.2.3
thinc_apple_ops==0.1.5
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
typeguard==4.2.1
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
visions==0.7.6
wandb==0.17.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wordcloud==1.9.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
ydata-profiling==4.8.3
zipp==3.18.2
zope.interface==6.4.post2
